rubocop修正_2022103001

+++ db/migrate/20210508072557_add_name_to_users.rb
@@ -4,7 +4,7 @@ class AddNameToUsers < ActiveRecord::Migration[6.1]
   def change
     add_column :users, :name, :string
     add_column :users, :role, :integer, default: 0, null: false
-    add_column :users, :organization_id, :integer, default: 1, null: false
+    add_column :users, :organization_id, :bigint, default: 1, null: false
     add_column :users, :is_valid, :boolean, default: true, null: false
   end
 end
