+++ db/migrate/20220821125509_create_folders.rb
@@ -2,7 +2,7 @@ class CreateFolders < ActiveRecord::Migration[6.1]
   def change
     create_table :folders do |t|
       t.string :name
-      t.integer :organization_id
+      t.bigint :organization_id
       t.integer :video_folder_id
 
       t.timestamps
